refactor: extract requirement review workbench
